+++ UC4-chaincode/src/main/java/de/upb/cs/uc4/chaincode/ApprovalContract.java
@@ -1,6 +1,7 @@
 package de.upb.cs.uc4.chaincode;
 
 import de.upb.cs.uc4.chaincode.error.LedgerAccessError;
+import de.upb.cs.uc4.chaincode.model.Approval;
 import de.upb.cs.uc4.chaincode.model.InvalidParameter;
 import de.upb.cs.uc4.chaincode.util.ApprovalContractUtil;
 import de.upb.cs.uc4.chaincode.util.GsonWrapper;
@@ -40,12 +41,14 @@ public class ApprovalContract extends ContractBase {
             return GsonWrapper.toJson(cUtil.getInternalError());
         }
 
-        String id = cUtil.getDraftId(ctx.getClientIdentity());
-        return cUtil.addApproval(stub, key, id);
+        Approval approval = new Approval()
+                .id(ctx.getClientIdentity().getId()) // use simple id for now
+                .type(ctx.getClientIdentity().getAttributeValue("hf.Type"));
+        return cUtil.addApproval(stub, key, approval);
     }
 
     @Transaction()
-    public String getApprovals(final Context ctx, final String contractName, final String transactionName, final String... params) {
+    public String getApprovals(final Context ctx, final String contractName, final String transactionName, String... params) {
         ChaincodeStub stub = ctx.getStub();
 
         ArrayList<InvalidParameter> invalidParams = cUtil.getErrorForInput(contractName, transactionName);
@@ -55,12 +58,15 @@ public class ApprovalContract extends ContractBase {
 
         String key;
         try {
+            if (params == null) {
+                params = new String[0];
+            }
             key = cUtil.getDraftKey(contractName, transactionName, params);
         } catch (NoSuchAlgorithmException e) {
             return GsonWrapper.toJson(cUtil.getInternalError());
         }
 
-        ArrayList<String> approvals;
+        ArrayList<Approval> approvals;
         try{
             approvals = cUtil.getState(stub, key);
         } catch(LedgerAccessError e) {